requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-filter==2.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
Markdown==3.1
nltk==3.4
Pillow==5.4.1
prompt-toolkit==2.0.9
pytz==2018.9
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
wcwidth==0.1.7
wit==5.1.0
